refactor: move xml processing to utility
