+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod config;
 pub mod core;
 pub mod error;
 pub mod sinks;
+pub mod xml_utils;
